Add functionality for authorization (for editing/deleting posts) on server.

+++ controllers/posts.js
@@ -17,9 +17,15 @@ exports.getPosts = (request, response, next) => {
 	Post.find().countDocuments()
 		.then(itemCount => {
 			totalPosts = itemCount;
-			return postQuery;
+			return postQuery.populate('creator');
 		})
-		.then(fetchedPosts => {
+		.then(result => {
+			const fetchedPosts = result.map(p => {
+				// remove the creator's (hashed) password from the responses!
+				const post = p._doc;
+				delete post.creator.password;
+				return post;
+			});
 			response.status(200).json({
 				message: 'success!',
 				posts: fetchedPosts,
@@ -34,6 +40,7 @@ exports.storePost = (request, response, next) => {
 		title: request.body.title,
 		content: request.body.content,
 		imagePath: constructImagePath(request.protocol, request.get('host'), request.file.filename),
+		creator: request.userData.userId,	// this was attached to the request in 'check-auth' middleware!
 	});
 	postToStore.save()
 		.then(savedPost => {
@@ -50,8 +57,14 @@ exports.deletePost = (request, response, next) => {
 	const idOfPostToDelete = request.params.postid;
 	const imagePath = request.query.imagePath;
 	const imageType = request.query.imageType;
-	Post.deleteOne({ _id: idOfPostToDelete })
+	Post.deleteOne({ _id: idOfPostToDelete, creator: request.userData.userId })
 		.then(result => {
+			if (result.deletedCount === 0) {
+				// This means that no posts were modified, meaning that the user is not the creator of this post!
+				const error = new Error('Not authorized to delete this post!');
+				error.statusCode = 401;
+				throw error;
+			}
 			if (imagePath && imageType) {
 				const imagePathToDelete = imagePath + '.' + imageType;
 				const fullImagePath = constructImagePath(request.protocol, request.get('host'), imagePathToDelete);
@@ -80,8 +93,14 @@ exports.updatePost = (request, response, next) => {
 		// If the request has a file, we have to delete the previously uploaded file for this post!
 		clearImage(request.body.imagePath);
 	}
-	Post.updateOne({ _id: idOfPostToUpdate }, updatedPost)
+	Post.updateOne({ _id: idOfPostToUpdate, creator: request.userData.userId }, updatedPost)
 		.then(result => {
+			if (result.nModified === 0) {
+				// This means that no posts were modified, meaning that the user is not the creator of this post!
+				const error = new Error('Not authorized to update this post!');
+				error.statusCode = 401;
+				throw error;
+			}
 			response.status(200).json({
 				message: 'success!',
 			});
